+++ doc/release/release_notes.rst
@@ -16,6 +16,10 @@ Release Notes
         Documentation specifies Python 2.6 instead of Python 2.7 as minimum
         interpreter version.
 
+    .. change:: fixed
+
+        Documentation does not reflect current dependencies.
+
     .. change:: changed
         :tags: session, component, locations, performance