+++ doc/release/release_notes.rst
@@ -11,6 +11,15 @@ Release Notes
 
 .. release:: Upcoming
 
+    .. change:: fixed
+
+        Documentation specifies Python 2.6 instead of Python 2.7 as minimum
+        interpreter version.
+
+    .. change:: fixed
+
+        Documentation does not reflect current dependencies.
+
     .. change:: changed
         :tags: session, component, locations, performance