+++ doc/release/release_notes.rst
@@ -17,6 +17,15 @@ Release Notes
         Increased performance of :meth:`~ftrack_api.query.QueryResult.first` and
         :meth:`~ftrack_api.query.QueryResult.one` by using new `limit` syntax.
 
+    .. change:: fixed
+
+        Documentation specifies Python 2.6 instead of Python 2.7 as minimum
+        interpreter version.
+
+    .. change:: fixed
+
+        Documentation does not reflect current dependencies.
+
     .. change:: changed
         :tags: session, component, locations, performance